+++ pkgs/artisanal_widgets/test/components/wizard_test.dart
@@ -1,7 +1,7 @@
 import 'package:artisanal/terminal.dart' as terminal_keys;
+import 'package:artisanal/testing.dart';
 import 'package:artisanal/tui.dart' show InterruptMsg, KeyMsg;
-import 'package:artisanal_widgets/artisanal_widgets.dart';
-import 'package:artisanal_widgets/testing.dart';
+import 'package:artisanal/widgets.dart';
 import 'package:test/test.dart';
 
 void main() {